requirements.txt
aiofiles==23.2.1
aiohttp==3.8.4       
aiosignal==1.3.1     
anyio==3.6.2
APScheduler==3.10.1  
asgiref==3.7.2       
asttokens==2.2.1     
async-generator==1.10
async-timeout==4.0.2 
attrs==21.4.0        
backcall==0.2.0      
bcrypt==4.1.2        
bfres==0.1.0
bidict==0.23.1       
binaryio==1.0.0      
blinker==1.6.2       
byml==2.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
cryptography==37.0.2
cycler==0.11.0
dateparser==1.1.7
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.13
discord.py==2.2.2
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.101.0
fastapi-utils==0.2.1
fastapi_amis_admin==0.6.3
fastapi_scheduler==0.0.13
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
flask-sock==0.7.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.34.4
frozenlist==1.3.3
greenlet==2.0.2
h11==0.13.0
idna==3.3
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.23.1
oead==1.2.6.post1
openpyxl==3.1.2
outcome==1.1.0
packaging==21.3
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.3.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.6
PyGithub==1.58.1
Pygments==2.15.1
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-engineio==4.9.0
python-multipart==0.0.6
python-socketio==5.11.1
pythonping==1.1.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==5.4.1
pyzmq==25.0.2
regex==2022.10.31
requests==2.28.2
selenium==4.1.5
simple-websocket==1.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.28
sqlalchemy2-stubs==0.0.2a35
sqlalchemy_database==0.1.0
stack-data==0.6.2
starlette==0.27.0
tornado==6.3.1
traitlets==5.9.0
trio==0.20.0
trio-websocket==0.9.2
typing==3.7.4.3
typing_extensions==4.10.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.9
uvicorn==0.21.1
wcwidth==0.2.6
Werkzeug==3.0.1
wrapt==1.15.0
wsproto==1.1.0
WTForms==3.1.2
xlrd==2.0.1
yarl==1.8.2